Fix incorrect array size specification
The count and start arrays are dimensioned to NC_MAX_DIMS even though the maximum size should be NDIMS which is set to 3; possibly, the maximum size could be 2 since only indices 0 and 1 are used to access both start and count.  I left it at NDIMS since that matches the number of items in the initialization and is consistent with other uses in the file.