Generate correct results from cmake-generated nc-config
A couple of the `has_XX` variables seem to have values of OFF or ON instead of `empty` or ON which causes incorrect output from the nc-config script since it is using `-z` to test whether the variable is non-empty.  This causes nc-config to incorrectly report that the library supports HDF5 or has the version 2 api.