Use correct hdf5 library in check_library_exists calls
Some of the check_library_exists calls were using a generic "hdf5" for the library to search.  This cannot always be found if the user-specified hdf5 library is not in the default search path.  If ${HDF5_C_LIBRARY} is used instead, then the library that will be used in the build will be searched instead.  These calls now match others that were already using that symbol.