Move setting of default HDF5_C_LIBRARY symbol
If HDF5_C_LIBRARY is not specified and can't be found, set to hdf5.

This replaces the current code which sets this symbol to hdf5 no matter what, but does it after some tests that use the symbol. 
Changed code only sets the symbol if not already set and moves it before the tests using the symbol.