The recent change that added #define _WINSOCKAPI_ in the latest config.h.cmake.in causes (all?/some?) versions of curl.h to throw a syntax error when compiling with Visual Studio.
I did a quick test by removing the #define _WINSOCKAPI_ from
config.h.cmake.in and the syntax error went away.

I do not know if the latest version of libcurl (7.57?) still has
this problem.

Ideally, curl.h should be fixed, but this will take time.

The next best fix (implemented in this pr) is to modify all
headers that have
#include <curl/curl.h>
to place
#undef _WINSOCKAPI_
before the
#include <curl/curl.h>