Move nels==0 check instead of eliminating
The previous change where the `nels==0` check was removed caused problems with nc_test.  
If the check is moved after the `NC_put_vara` call, then it will avoid the parallel hang problem for stride=1, but will still pass all tests in `nc_test`.  Note that this is somewhat of a kluge since there will still be a parallel hang if nels==0 and stride is > 1 and the code falls into the odometer section.