re: e-support FLB-437472
Note: base branch for this is currently master rather than v4.5.

The storedump function is apparently failing on a very simple .dodsrc:
> HTTP.COOKIEJAR=/home/aswin/.rda_cookies
> HTTP.NETRC=/home/aswin/.netrc

It turns out that the problem is that I was forgetting to test
that the host field of a triple was null before trying to get
the length of the host field. Searched and fixed a couple of
similar situations in libdap4 and oc2.

Added a test of sorts by creating a more complex fake .dodsrc
file that would fail the old code.