Fix (I think) the --has-dap and --has-dap4 output
The `has_dap4` define was missing trailing @.  I also added tests similar to other code to result in lowercase yes no for `--has_dap4`.

I'm not as sure about the `--has-dap` changes.  One place was outputting `$has_dap` which is never defined.  I think that the intent is that `--has-dap` returns the same as `--has-dap2` so made changes that did that.  If that isn't the intent, then let me know and I will back out those changes and only leave the dap4-related change.