Remove useless checks
The requirement that `mode` is equal to `NC_MPIIO` or `NC_MPIPOSIX` or `NC_PNETCDF` is checked, but is not ever used after this point.  `NC_MPIIO` is now the same as `NC_PNETCDF` and `NC_MPIPOSIX` is deprecated.

Once the execution gets down into `NC_open`, the mode has `NC_MPIIO` set automatically (`if(useparallel) flags |= NC_MPIIO;`) where `flags` is the same as `mode`.