Fix unused variable compiler warning
Fix unused variable warning for `printid`.   Looks like it was not substituted in all places where it should have been and was only used inside an ifdef so was unused if `LOGGING` was false.  Add use in `fprintf` line following the ifdef block.

Minor change, but compilation output is so clean that a little wart sticks out...