Remove compiler warning
If `ENABLE_SET_LOG_LEVEL` is not defined, then `nc_set_log_level()` is removed via a macro such that `nc_set_log_level(val)` will be replaced by the C preprocessor to be nothing.  This works fine in the code, but in the `netcdf.h` include file, it results in an empty declaration of `EXTERNL int;` which causes the compiler warning: `TPL/netcdf/netcdf-c/include/netcdf.h:1751: warning: useless type name in empty declaration`.

This change removes the declaration of `nc_set_log_level()` if `ENABLE_SET_LOG_LEVEL` is not defined.