Eliminate compiler warning
The `typedef struct NC_Dispatch NC_Dispatch;` declaration is in both netcdf.h and in ncdispatch.h.  Even though it is the same declaration, there are compilers which see this as an error and the compilation will fail:
```
In file included from /netcdf-c/include/nc4dispatch.h:15,
                 from /netcdf-c/libhdf5/hdf5attr.c:16:
/netcdf-c/include/ncdispatch.h:111: error: redefinition of typedef `NC_Dispatch`
/netcdf/netcdf-c/include/netcdf.h:515: note: previous declaration of `NC_Dispatch` was here
```
This is with a very old version of gcc-4.4.7 which isn't used much anymore, but this is the default gcc compiler on RHEL6 so it might be worth fixing.