Eliminate double-free in realloc'd memory
If `memio->memory` is freed, then there is a double-free error since that pointer was either already freed in the realloc call, or it is the same pointer as newmen and should not be freed.   The documentation of realloc(3) states:
```
If the area pointed to was moved, a free(ptr) is done.
```

I'm not sure why I haven't hit this before, but I had a couple of my tests start throwing `double-free` errors and valgrind confirmed that the pointer had already been freed at the point of the realloc call.