Modification in support of eventual Zarr support.
This is a set of changes to the way that URLs are handled
as paths. This is in support of eventual Zarr support.
The primary difference is to extract io handler information
from the url. This will be used inside dispatchers that support
multiple IO access methods (like file versus Zarr) to figure out
what internal IO handler to use.

Specific changes:
1. move url analyzer code from dfile.c to durlmodel.c
2. add code to walk the .daprc triples
3. NC_HDF5_finalize was not being properly called by nc_finalize().
4. Fix minor bug in ncgen3.l
5. fix memory leak in nc4info.c
6. replace protocol= fragment tag with a more general mode= tag.